Added method to User model to return groups a user is (non-)member of
For this we needed to query the `group_members` table in the user-backend. To
disconnect the coupling between a certain table and a backend-file, the setup
functions were moved to one setup function in the backend's `__init__.py` file.

The setup is now performed in a single function, callable through
`models.backendSetup()`.



git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/branches/sql@906 9dfea958-5e0b-0410-aa7e-b09452dc4efc