Create all remaining tables (repositories, notifications and permissions_*)
Removing groups and users are now splitted up into multiple functions which
remove different aspects of that user/group. This way, the contract is even
more clear to the backend designer: it has to implement a few extra functions,
but can now no longer forget to remove certain aspects of a user / group.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/branches/sql@913 9dfea958-5e0b-0410-aa7e-b09452dc4efc