Load settings lazily. Refs #170
Only import settings when a setting is requested for the first time. Bootstrap
now also provides a setSettings function to override the settings with (for
example) a mock-object.

To allow such a mock-settings-object to be passed to a backend, we added a
backendInit which calls plugins.backends.<backend>.init with the passed
settings-object.

This commit also introduces the backendTearDown (which calls
plugins.backends.<backend>.teardown), which can be used for closing a database
or final flushing of files etc.



git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/branches/sql@930 9dfea958-5e0b-0410-aa7e-b09452dc4efc