Unittests for models now use new settings setup. Refs #170
A new backend, 'mock' was added, which keeps an in-memory "db" of users and
groups. This db is emptied on every test.

Furthermore, the UserTests-suite now uses the new model-structure.
Notifications still fail since we have not implemented them in the models yet.

Furthermore, I disabled the RepositoryTests-suite since we have no (new-style)
repository model yet.

For completion of #170, we still need backend unittests.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/branches/sql@931 9dfea958-5e0b-0410-aa7e-b09452dc4efc