Added tests for the Group model
For the Group tests to pass, the mock-backend needed to conform more to the
contract.

Furthermore, the model-tests are now split up in separate files in the
models.tests-package. A small change to runtests.py was needed to facilitate
this extra package by introducing the suite() function. This function returns a
TestSuite which can be added by the runtest-script.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/branches/sql@937 9dfea958-5e0b-0410-aa7e-b09452dc4efc