Implemented system-hooks and plugin-hooks system
Plugins can now add a hooks.py with a dictionary hooks containing mapping an
event to a list of functions.

Hooks not related to a plugin are placed in the dictionary hooks in
submin.hooks.system_hooks.

User-hooks are still placed in the directory hooks in their environment as
described in GenericHooks.

Refs #191



git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1177 9dfea958-5e0b-0410-aa7e-b09452dc4efc