Fixed typo, cached import to prevent error messages.
Two problems were found, one hiding the other. Found because code sometimes
complained about the following error:

    Exception AttributeError: "'module' object has no attribute
    'StorageException'"

Subminadmin was checking for a StorageException, but this class doesn't exist,
it is StorageError. After that typo was fixed, the second problem was
uncovered, as shown by this error message:

    Exception RuntimeError: RuntimeError('sys.meta_path must be a list of
    import hooks',)

For some reason, the import failed during closing, but not during opening.
This was 'solved' by caching the imported module and reusing that when
closing.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1360 9dfea958-5e0b-0410-aa7e-b09452dc4efc