Add Trac CGI support
Apparently, mod_python is not under active development. Trac recommends to not
use mod_python anymore. CGI support also enables us to have a default option
for quickstart that is more likely to work out-of-the-box.

Because trac.cgi (and friends) are licensed differently and because the
contents of the scripts can change with new trac versions (improved versions),
we generate the scripts. Because we need a working trac environment for that
(sigh), we generate that one too.

Additionally, when using 'apacheconf create all' the template is now optional.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1379 9dfea958-5e0b-0410-aa7e-b09452dc4efc