Used CSS sprites for the arrow keys as well
This involved some refactoring:
1) CSS sprites can't scale images and we used image
   scaling. Because we fix the font-size to 12px and
   the arrows to '1em', the images are now fixed to
   12px (width/height) as well.
2) Of course the images are now all in one file
3) Because we use div's instead of img's, a lot of
   templates and javascript had to be changed

Overall I'm quite happy with the result. Again less
HTTP requests and I think the javascript is now cleaner.

There is only one thing that I don't like yet and that is
that for every collapsable collection (sidebar, users,
repositories) we have to define a CSS class. This can
go away once we rewrite the collapsables code to use
different classnames instead of one.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1386 9dfea958-5e0b-0410-aa7e-b09452dc4efc