Encoding improvements (refs #278)
Encoding is not fully handled centrally in Submin. This
commit aims to improve that. Instead of converting to
Unicode for each post.get(), all post variables are now
converted to Unicode by default. Also, the path objects
are encoded to Unicode by default. This saves a lot of
encoding in views (where it should not be done), especially
considering a lot of Javascript implementations use
'%uXXXX' (apparently not a 'real' standard, but a de facto
standard nonetheless), but the Python libraries (e.g. cgi's
parse_qs) do not convert these characters at all.

The 'User' model now has a __unicode__() method and the
__str__() method encodes to 'utf-8' explicitly.

The WSGI/CGI request objects are refactored, because they
both had a WSGI/CGI Field Storage class, that differed only
in 'keep_blank_values=True' (which should have been on in
both, to keep the behaviour the same).

All in all, more centralized encoding handling (on input).
Encoding improvements are not complete yet, some cases on
output are not handled correctly yet.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1396 9dfea958-5e0b-0410-aa7e-b09452dc4efc