Use JSON in favour of cPickle, so we can handle Unicode (fixes #278)
Actually, we also don't put complete User objects in
the session anymore, only the name, admin state and
authentication state. This makes the session a little
bit smaller and it allows cross-platform session
access (not that it is needed, but hey). We were forced
to because JSON doesn't allow all classes to be 
JSON-ified

Actually, I think it's nicer because now there are less
User objects created, only when they are really needed.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1407 9dfea958-5e0b-0410-aa7e-b09452dc4efc