Added fallback to cPickle
Since json is included from Python 2.6, using json
would exclude Python <2.6 users. This fallback allows
those installations to work, with the restriction that
non-ascii characters cannot be used.

Another solution would be to include simplejson, but
since it is (yet another) different Licence, and I'm
not a lawyer, I choose to go with a solution that
uses the standard libraries instead.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1408 9dfea958-5e0b-0410-aa7e-b09452dc4efc