Changed URL to redirect to when logging out
For some reason, the user was redirected to
whatever was in 'redirected_from' after the user
logged out. This makes no sense, as this page is
most likely not accessible (because of the logout)
and will result in a redirect to the login page
with the 'redirected_from' set to that URL. After
logging in as a valid user, this user will be
redirected to that URL.

Technically strong, but not the right thing to do.
This could cause the following weird effect:
 1) user logs in as normal user N (going to its own
    profile page).
 2) user N logs out (setting redirected_from to user's
    profile page)
 3) admin user A logs in, gets redirected to user N's
    profile page.

To fix this, no looking at the redirected_from anymore,
just redirect to submin_base_url.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1409 9dfea958-5e0b-0410-aa7e-b09452dc4efc