Don't redirect to login after login
When the first page is <submin>/login, and the user is not logged in, this
url will be used as 'redirected_from'. After logging in, the user will be
redirected back to the login page, which is counterintuitive. The
workaround is to remove the '/login' part, but this commit checks if the
user would be redirected to <submin>/login and if so, redirects to home
instead.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1546 9dfea958-5e0b-0410-aa7e-b09452dc4efc