Hint to the user that she needs to set permissions (refs #314)
After creating a Trac repository, the user might not be aware
that Submin does not manage repositories automatically. This
commit introduces a hint that is shown once after creation of
the Trac environment, informing the user that she is now
TRAC_ADMIN and that Submin is not managing Trac permissions.

Together with the link introduced in [1545] (direct link to
Trac permissions manager), this makes it much easier for the
user to setup Trac correctly.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1547 9dfea958-5e0b-0410-aa7e-b09452dc4efc