Added 'chown-socket' to ini file
I needed this option otherwise I would get an error message concerning bind:

  bind(): Permission denied [core/socket.c line 185]

My guess is that the socket is not read/writable by the user:group, so the
uwsgi server dies (and removes the socket). This is maybe due to a new
version of uwsgi, since this error occured after an upgrade.


git-svn-id: https://ssl.supermind.nl/collab/svn/submin/submin/trunk@1554 9dfea958-5e0b-0410-aa7e-b09452dc4efc