Get apache module list via server-info (refs #330)
In some cases, it's not possible to get a list of loaded apache modules via
apachectl. This is the case when permissions are too restrictive for the
apache user to properly parse the config. Examples are restrictive log
directories, or SSL configurations.

The old method is still tried first, but if the 'apachectl' variants fail, it
is attempted to get the list via the /server-info?list URL. This is only
possible if the 'info' module is loaded, and configured for use.

The URL is by default 'http://localhost/server-info?list', but can be
specified by setting the 'apache_server_info_url' config with submin2-admin.