Versioning of generated git hooks (refs #332)
Generated git hooks are now versioned. This makes it easier to detect
if hooks are up-to-date or not. This can be verified in the Diagnostics
page. This work is preparation for a change in the post-receive hook to
make sure that git-multimail is only called when there are actually email
recipients.