Only invoke hook if mails need to be send (refs #332)
Check git config to see if any mails are going to be send. If not, simply
do not run the git-multimail hook, since it is not necessary anyway.

This prevents confusing error messages when a repository has no no users with
commit messages enabled. This can occur after a newly created repository, or
if all users have disabled commit messages.