Added wrapper for os.makedirs
I found myself making the same code over and over again: os.makedirs with
a try/except block around to catch if the leaf directory already exists.

This wrapper creates the directory if it doesn't exist, but doesn't throw
an exception if it already does exist. It does throw exceptions for everything
else, so you can still use try/except for other exceptions.

Also added a unittest for this code.