Enable post-receive multiplexer (refs #341)
There used to be only one post-receive script: the email script. But for
Trac Sync to work, we need to run two post-receive scripts. This is
difficult because the post-receive script read standard input, thus the
first script will deplete all the data before the second (etc.) script
had a change to run.

This changeset adds support to run multiple scripts. The API between
"submin2" and "submin2-admin . git admin" has changed. Instead of the
'post-receive-hook' command, there are now two commands called
'commit-email-hook' and 'trac-sync-hook'.

Instead of one post-receive-hook script, the new layout in the git
'hooks' directory will now be:

 [...]
 post-receive-hook -> [...]/hook-mux # the multiplexing script
 post-receive.d/                     # directory that contains hooks to run
 post-receive.d/001-commit-email.hook # exists iff commit emails are enabled
 post-receive.d/002-trac-sync.hook   # exists iff trac sync is enabled
 [...]

This hook-mux can be used for other hooks that receive data on stdin.

The diagnostics is not updated yet, so for the moment, this will give a
warning that new post-receive-hooks are not up to date (but the old ones
will be shown as up-to-date).