Initial support for trac/git sync hook (refs #341)
Unfortunately, we had to make a lot of changes, because the hook
is run as the git user, while the trac environment is write-only
for the web user. Both need write access, so this is not trivial.

There are several options to continue:
 * run trac as the same user as git/ssh (easy, but not secure)
 * run something setuid (not easy, and difficult to get secure,
   especially with scripts).
 * do something through sudo (not that ugly, but creates a dependency
   on sudo).
 * let the post-receive write something in the database, and connect to
   the webserver to trigger an action. You can either wait for that
   action, or shoot-and-forget. (this will add another layer of
   complexity, and points of failure)

I have chosen to implement the last option, since I thought it was the
least ugly one. Putting the data in the database in the mean time will
help with robustness, so I'm confident this will work smoothly.

The status is that commit messages are stored in the database, and they
are automatically synced, but the machine of the git user should be in the
IP ACL. If it is not in the IP ACL, it will simply dump the error page as
HTML on your git client output. This is not nice.