Allow 'False' in template
The handling of booleans was not symmetrical in the template module.
When a True boolean value was encountered, it would show up as a string
'True'. However, when this boolean value was False, this resulted in the
string value '' (instead of 'False').

This is now changed. Both 'True' as 'False' show up in the final string,
and the value None still results in '', as before.