Clearer error message when IP denied (refs #341)
Whenever someone pushes changesets, the information about these changesets
are temporarily saved in the database and a URL is retrieved so the web
user can update Trac.

This URL is protected to prevent abuse, so an IP-address has to be
configured. If this is not configured correctly, the git client will
receive an error message. This error message is now clearer.

Instead of letting the dispatcher take care of the exception that is
raised in this case, we only set the acl_required decorator on the
trac-sync handler so we can catch the exception and return an
XMLTemplateResponse instead. This is much less output, and allows the
other side to interpret this programmatically. That last part is not done
yet, but at least it is more readable than a backtrace.