New diagnostic: trac hook access (refs #341)
There is a new diagnostics, to check if submin has access to the sync hook.
This check will connect to submin (itself!) to see if the IP address has
access.

Because it will connect to itself, this check will not work for web servers
where there can only be one single connection simultaneously (our development
server for example!). In those cases, it will timeout.

A new hook is introduced for this, but I don't like this, so I'm going to
rewrite this.