Code cleanup (refs #341)
Instead of checking a separate 'test-acl' hook, we now check the hook
directly. This is possible because the XML now returns an 'inacl' attribute
if the IP address is in the ACL (so we can programmatically check if we are).

Also the internal REST API has changed; the order of directories is different.
Instead of /<vcs_type>/<repo>/<hook_type>, a request is now made by:
/<hook_type>/<vcs_type>/<repo>. This makes it possible to have different
parameters for each hook (I should've done this in the first place anyway).

There were other code cleanups as well.