Improve diagnostics and trac-sync robustness (refs #341)
Trac sync uses a different default if 'acl_hook' is not set. It now uses not
only localhost IP-addresses, it also tries to resolve the hostname that
belongs to the vhost of submin and uses those IP-address. Since the default
of 'acl_hook' is unset, this should work for most people.

Diagnostics will check if the acl_hook is set, and will -warn- if it's not
set, but it will also check if trac-sync has access. If the variable is not
set, it will suggest the same setting as it will use by default. By
explicitly setting this to the IP-addresses means that there will be no DNS
lookups at every (git?) push.

Also changed diagnostics to give commands more emphasis. They now have their
own box, so it's clear what you need to copy past if you want to run commands.