Don't forget about the old style commit hook (refs #341)
During tests we found the old email commit hooks were lost during an
'submin2-admin ... git hook update [...]' operation. This was due to the
old hook not being recognised, so when the post-receive-mux was put into
place, the commit-email hook was not destroyed.

This fix will correctly show old-style hooks and preserve them on
upgrades.