Call hook-mux like update script (refs #341)
The symlink to hook-mux didn't work in all cases, because the executable bit
was not always set. This is because setuptools does not enforce this (unless
your script should be moved to /bin or /usr/bin, which is not what we want).

To prevent problems, we now call the hook-mux script for post-receive like
the update script. This allows us to update the hook-mux script (or the update
script) by just installing a new version of submin, while at the same time
ensuring that it is always executable (submin controls permission bits in
the repository, but not in the install directory).

To debug all of this, diagnostics is now more verbose as to WHY hooks are
outdated.

Because similar code was needed in the post_receive_hook.py module, the
git hook code was refactored as well. Big code changes for such a small
change in functionality, but I think the code is a bit cleaner now.