And this is why you should test (refs #341)
So, I forgot that the hook-mux script worked by checking how it was
linked to the hook-mux script and uses that as hook name. Of course
it doesn't work if you just call it.

I've changed the hook-mux script to accept an argument, which is
interpreted as the hook name (if no argument is found, it will try
to get it via the old way).

Then I found some .git extension inconsistensies (refs #296), and
worked around it for now (since we haven't decided yet what we're
going to do).