Add some files to dummy packages (refs #343)
We need to add some files to the dummy packages in order to prevent a
warning message that says these packages have been replaced completely.
This is due to the fact that debian notices that all paths of a
previously installed packages have been overwritten.

We now add some README files explaining this to each package.