Export users to htpasswd
When someone doesn't use Apache and its authn_dbd module for
authentication for trac (but, say, tracd behind an nginx proxy), we need
an htpasswd file with all users and password-hashes.

This change implements an export to a file specified in the
htpasswd_file option. Since the passwords are salted md5-hashes, tracd
and nginx can parse these already. We don't need to have access to the
original, plain-text passwords for this.