Add new git_multimail.py
The new version of git-multimail has some changes in the post-receive hook.
The config parameter is now a keyword instead of a positional argument. Also
the way of selecting a mailer is different.

This version of git-multimail adds support for artificially increasing the
date (by one second each commit mail inside one push) to improve sorting
in mail clients.

The version of git-multimail was from: Tue Feb 11 06:48:40 2014 +0100

Submin's post-receive hook template is now updated, this means that existing
installations need to upgrade their template. A next commit will include
support for fixing migration issues.

This also fixes an issue where the email subject was: '[Submin] [<project>]'.
It is now simply '[<project>]'.