Update git_multimail.py for old post-receive hook
These changes are not present in the upstream version of git-multimail,
these are just compatibility fixes with the previous version of
git-multimail that was bundled with Submin.

In the future people should update their template so that we can install
the next upstream version without changes, but that means they should
check their diagnostics after an upgrade. For now, we leave our changes
in, so both versions will work.

The changes are:
 * In the old version, the headers template included an empty line, to
   signify end of headers. The new version doesn't have this, so fails
   if a newline is present at the end.
   Solution: catch Exception and continue
 * We called SendMailer with the envelope sender as the first positional
   (non-keyword) argument. The new version expects command first.
   Solution: swap arguments
 * The old version had a member for the envelope sender
   Solution: add getter property for 'sender'
 * The old version had one positional argument for the GenericEnvironment,
   the new one only supports keyword arguments.
   Solution: override __init__, pass the first positional argument as a
   keyword argument