Improve upgrades to new split-package (refs #343)
Automatic upgrades didn't go well because submin2-core has files that
also exist in the old version of the submin2 package (everything moved
from submin2 to submin2-core, while the rest are just dependencies).

Introducing a conflict with submin2 << 2.1.3 will fix this. It needs to
be bumped because nightly packages have multiple hyphens, confusing the
version numbering. This is a packaging bug.