Fix the SSH host questions (refs #352)
It seems submin didn't guess hostnames at all, it just defaulted to
localhost for the public hostname and took the new value for the
internal one, which made it look like submin reversed the two.

Now guess the hostname in a similar way as in 'initenv', and set the
internal hostname to 'localhost' by default. Much better!