Re-introduced output in CalledProcessError
In python 2.6, there was no 'output' attribute for the CalledProcessError
exception; there was no 'check_output()' there either. Because of this,
a wrapper was written in submin/common/execute.py, but this wrapper didn't
add the output if the code was run by python 2.6.

The output was removed in a previous commit (7271528), because it didn't work
in python 2.6. This change will reintroduce the output simply by attaching
it to the exception after it has been created.