Do not rely on SUBMIN_ENV, set it ourselves.
Running the tests was impossible without setting SUBMIN_ENV, which is
an unwanted dependency. This is because c_apacheconf needs to include
'options' to get the basedir for templates, but we don't actually need
it for these tests (yet!).

So for now, we create a directory, put settings.py with the content in it,
but we don't actually need to create an SQLite db, because it's not accessed
at all. Also, we set the environment variable to the created directory.

This directory is basically a submin-env mock, but on disk. Newer tests may
need more, but we can always add files/dirs to the mock, or possibly just
run s-a initenv (but that will trigger an email to be sent, so that might not
be a nice option).