Use http_vhost instead of guessing schema in redirects
I was still noticing an incorrect redirect to the login view. This was
caused by a wrong guess of the url scheme in the Redirect response,
which in turn was (probably) caused by the fact that uWSGI does not know
what to use for the HTTPS environment variable (since it runs as a
unix-socket).

This change moves away from guessing the schema based on the environment
variable, and uses the `http_vhost' option instead. On my server, this
resolves the issue and I am now correctly redirected to the https
version of the login page.

Fixes and refs #334.