Added a Mock VCS type
When adding features or testing, it can be a drag to install support for
either subversion or git (because you don't want to install apache/svn
or you don't want to create a new user, or any other reason). We now
have a Mock VCS type, that just creates a directory and does not do
anything else. Enable it in your environment by adding mock to
vcs_plugins.