Fix coverage and unittesting
Trac tests didn't work correctly. This wasn't found because they
were actually never run. Also coverage had some problems.

Coverage now includes -all- submin files, so coverage goes down quite
a bit (this is good, so we can see what kinds of tests we need for minimal
coverage).

The runtest.py didn't include the submin path correctly (didn't work in all
cases), so system files were tested instead of working dir files. This is now
fixed.

Files to be tested can now be reliably be passed on the command line. It is
not very efficient, but it is simple and fast enough for the amount files
in the submin directory.