Moved vcs-specific code to plugins (fixes #296)
By moving all vcs-specific code to the relevant plugins, there are less
hacks, resulting in cleaner code. Also, this gave us the opportunity to
handle the '.git' extension more consistently, hence the closing of the
ticket.

This change requires a database upgrade, making the database more
consistent with the code as well (no more '.git' extensions where the
repository is stored).

Generating a URL to access the VCS is now also done in the vcs plugin.

Some typos and style changes were fixed while cleaning up the code.