Make sure we use the dev submin modules
If sys.path is not set, and the user forgot to source dev/scripts/env, and
if submin is installed on the system, there will be no error, but the
bin/submin2-admin script will silently use the system modules instead of the
working directory modules.

We force the sys.path to the working directory so this will not happen.