Code cleanup (refs #362)
I was annoyed by the inconsistencies and roundabout way of doing things, so
I cleaned up the code to make it more consistent and straightforward.

The 'Trac' class has disappeared because it was basically a check if the trac
environment directory exists.

Handling missing config variables is now more plugin-specific (no
vcs-specific checks in the repository view)

The naming of things in the trac module has been changed; instead of
trac.tracCreateEnv(), trac.create().

Added missing functions in the mock vcs plugin (apparently not tested).