Create repository: select first VCS type (fixes #347)
We (ab)use the special 'iindex' variable to 'check' only the first radio
button. The order can be specified by changing the order in the settings.

For example, to have git as default, set it to:

submin2-admin /var/lib/submin config set vcs_plugins git,svn

To set svn as default, set the value to 'svn,git'.