Help prevent conflicting subversion configuration
Some distributions have default configuration for sbuversion that conflicst
with the apache configuration files that are generated by Submin.

This patch introduces an explicit warning in the generated SVN file and
instructs administrators to read this warning after generated the files
(closes #336).

Also some wording is changed to (hopefully) make it easier to read.