Diagnostics: add Trac environment checks
The new diagnostics checks are:
 * check if a trac env points to an existing repository
 * check if all necessary modules are enabled

Modules that should be enabled include:
 * vcs-specific connectors (tracopt.versioncontrol.<vcs>.*)
 * commit updater (tracopt.ticket.commit_updater.*)

If that last one is not enabled, comments will not be added after a
commit/push (refs #371).

The options will be checked including wildcards, so if for example
tracopt.versioncontrol.<vcs>.<whatever> is not present, but
tracopt.versioncontrol.<vsc>.* is enabled, the option will not be
reported as missing.