Diagnostics: add another Trac config check (fixes #371)
Added another check: commit_ticket_update_check_perms = false. This
should be set to false if you use git.

In order to add these test to the previous ones (they all tested
in the 'components' section, while this new one is in the 'ticket'
section), we changed the datatypes a little bit. Also, for each
option, we show if it is fatal or not, so the administrator can
decide better to ignore or follow our recommendation.

I think it's also better to say 'OK', 'Warn' and 'Fail' instead of
'Yes', 'Warn' and 'No', because we can use it in other parts as well
and it is easier to interpret (no need to read the text in front).