Added TOC to diagnostics (fixes #358)
Because adding the TOC changed the way everything looked, I've updated
many appearance things as well. The result is, I think, much cleaner
when there are no failures/warnings and when there are many failures or
warnings, the more clear separation of sections and their options makes
it easier to see what belongs to what.