Provide links to trac.ini as well (refs #371)
Because of the update in appearance, we slightly changed the warnings and
failures of the missing configuration. The text that was shown for each
environment is now shown once at the bottom of the list, while a link to
the trac.ini (copy+paste only) and the Trac Admin panel (href) is provided
for each problematic environment.

To better distinguish the preformatted text from the rest, the diag_pre
CSS class is now subtly hilighted (off-white background color + border).