Fix sidebar/content separate scrolling
Scrolling the sidebar would scroll the content div as well, if the
sidebar finished scrolling and you would try to scroll further.

By forcing both elements as 'fixed' and setting the bottom of the
content div to 0, this will treat both elements the same; fixed
position, but because of the overflow property, scrollable.