Fix for fix for sidebar scrolling
In [df8b0148c08003b2830245ba31c05c542e522848], a change was introduced
to fix the sidebar scrolling, but the width of the content wasn't correct.

In some browsers, the content div was very small, so now we explictly
set right to 0px and set some padding to allow for scrollbar overlap.