Catch exception during initenv (fixes #368)
When no MTA is installed, or no /usr/sbin/sendmail executable exists,
submin does not know how to send an e-mail. Instead of bailing out of
the installation we catch the Exception and print a warning and
continue as much as we can.

The user can recover by requesting a password reset for 'admin'.

For the record, whenever this happens, the following commands have
to be executed to finish the installation:
 * submin2-admin <env> upgrade hooks no-fix-unixperms
 * submin2-admin <env> unixperms fix
 * submin2-admin <env> apacheconf create all
 * submin2-admin <env> trac init # (optional)