Change color of verify password notitication
The color of the password notification was green, this is counterintuitive
for users as it marks something that went well. However, after typing in the
password, some action has to be taken.

Now the color shows the color of the other hints (yellow). Also, the message
does not disappear anymore, you have to click it, or finish the password
change (after it is finished, a new message shows up, which will disappear
automatically). This refs #377.