Move decision to warn/fail to backend
The template is not the right place to decide to show a warning or
failure, leave it to the modules themselves. This will make it possible
to automatically set group (svn, git, email, trac) labels to warn if
only warnings were found and set it to fail if at least one failure
was found.

The new way of setting results is to also set a label next to the normal
variable. For example, 'trac_acl_hook' is set to True or False and if it
is False, 'trac_acl_hook_label' will be set to 'warn' (otherwise 'ok').

The decision to set it to 'warn' or 'fail' is done in the helper method
add_labels(). See doc string for more information.