Hint when incorrect usage is detected (refs #383)
This is one of the most common misusage of submin2-admin for new users.
The user is unaware that an environment has te be supplied (even for
help... this is very consistent, but not intuitive).

For now give a warning (and wait two seconds so the warning can be read)
and then do what the user intented.

This blocks the user from doing actions on an environment called 'help'
that is in the current directory. If you really want to do that, you can by
passing the full path (or ./help) instead of 'help'.