Fix apache missing module diagnostics (fixes #390)
Treat 'apachectl'/'server-status URL loading' failing differently from 'missing
modules'. For the latter, show a list of required/missing apache modules so
sysadmin can load the right modules. For the former, show hints on how to solve
these problems.

Also fixed an CSS issue that went wrong while refactoring (diag_no was
renamed to diag_fail, but one was skipped when converting this, because this
was an exception)