Fix display errors in Safari
In Safari (OS X Mavericks as well as iOS 7) the icons were not visible. This
was because the 'addicons' were defined in the 'sidebar' div, but the sidebar
div was (intentionally) a little bit too small in height, to leave space for
the addicons bar. This was done to make scrolling correct.

Chrome and firefox rendered the addicons correctly, but Safari did not.
Changing the overflow property to 'visible' did show the addicons bar, but
destroyed the scrolling functionality. I'm not sure which implementations is
correct.

Anyway, the solution is to move the 'addicons' bar outside of sidebar, so it
is visible in all browsers and is not dependent on the overflow property.