Template: Added suppress_newline modifier
If a command is prefixed with '@', the newline directly following the
command node is suppressed. This means that the command should be followed
by a textnode that starts with a '\n' character for it to have any effect.

The reason for this feature is to allow for newlines in the template for
readability of the template, but no newlines in the output (also for
readability). This is not that necessary for HTML templates, but it is for
plain text templates.