Nicer formatting of diagnostics CLI (refs #393)
Using the new 'suppress newline' modifier of the templates, we increase both
the readability of the template as well as the output. The newlines are
suppressed on all 'test' and 'else' commands, even though it might not be
necessary for all of them. In any case it's better to be consistent.