Better handling of out-of-env git_dir (fixes #406)
When the git_dir is outside of the submin environment (default:
/var/lib/submin), the subdirs could be created with incorrect permissions.

If for example git_dir is set to /var/repos/git, and the /var/repos/git
directory does not exist yet, it was sometimes created as root:root (ok)
with permissions rwxr-x-- (not ok, since neither the git user nor the web
user can then access it).

A new function is added that checks which directories need to be made and
only for those directories, creates them and sets permissions to a sane
default. If it already exists, it is up to the sysadmin to have sane
permissions.

The git update command will check if it has access to write a file and
will complain if it can not.