Add checks for git_dir permissions (fixes #408)
When the permissions for the git_dir are wrong, it might be that all git
commands do work, but the repositories are not listed. In this and other
cases the Diagnostics will complain.

This is mostly necessary for when the git_dir is outside the submin env. The
problem during 'git init' is already fixed (refs #406), but for the case that
'git init' was run with a version before this one, diagnostics will now show
which directories are wrong and how to fix them.

The 'unixperms fix' command does not complain about this, but I think that's
not a good idea since it is used in other commands as well (sometimes used by
scripts), so I'm not sure what will happen if I suddenly add warnings there.

In any case, the Diagnostics will show the incorrect permissions, so if
people see problems, they should run diagnostics.