Improve submin2-admin initenv questions (refs #355)
The 'initenv' command first asks which features to enable. The questions after
that will depend on which features are enabled. So if one chooses to disable
trac, the trac dir will not be asked. In that case, also the trac deploy will
not be done.

This is also the first time that nginx is enabled by default (only the
creation of config files, like with apache). The default is to enable both
apache as well as nginx.