Diagnostics: test if we can send e-mail (closes #397)
This testing is not waterproof, since the only way to really test is to send
an e-mail and we do not do that. It might help with some basic problems like
bad username/password pair or smtp_hostname is set to a server that doesn't
listen on standard SMTP ports.