Add news about breaking installs (fixes #414)
People that have apt-listchanges installed will now get a warning message
explaining what the problem is and how to fix it.

I have tested if the version matters (because nightly builds have extra data
in the version), but putting an older version in the NEWS seems to work fine;
I have tried upgrading from 2.2.1-1 to 2.2.2-0 with '2.2.1-3' (nonexistent)
in the NEWS file. This worked because I saw the message with apt-listchanges.